['libs' pages] added 'view mode' selector & 'sort' menu for 'my libs, shared with me, shared with all, group, department' pages
- besides that, fixed bugs & improved many lines of code
