+++ frontend/src/components/cur-dir-path/dir-tool.js
@@ -7,7 +7,7 @@ import TextTranslation from '../../utils/text-translation';
 import SeahubPopover from '../common/seahub-popover';
 import ListTagPopover from '../popover/list-tag-popover';
 import ViewModes from '../../components/view-modes';
-import ReposSortMenu from '../../components/repos-sort-menu';
+import SortMenu from '../../components/sort-menu';
 import MetadataViewToolBar from '../../metadata/components/view-toolbar';
 import { PRIVATE_FILE_TYPE } from '../../constants';
 import { DIRENT_DETAIL_MODE } from '../dir-view-mode/constants';
@@ -36,13 +36,6 @@ class DirTool extends React.Component {
       isRepoTagDialogOpen: false,
       isDropdownMenuOpen: false,
     };
-
-    this.sortOptions = [
-      { value: 'name-asc', text: gettext('By name ascending') },
-      { value: 'name-desc', text: gettext('By name descending') },
-      { value: 'time-asc', text: gettext('By time ascending') },
-      { value: 'time-desc', text: gettext('By time descending') }
-    ];
   }
 
   toggleDropdownMenu = () => {
@@ -111,13 +104,6 @@ class DirTool extends React.Component {
     const isFileExtended = currentPath.startsWith('/' + PRIVATE_FILE_TYPE.FILE_EXTENDED_PROPERTIES + '/');
     const isTagView = currentPath.startsWith('/' + PRIVATE_FILE_TYPE.TAGS_PROPERTIES + '/');
 
-    const sortOptions = this.sortOptions.map(item => {
-      return {
-        ...item,
-        isSelected: item.value === `${sortBy}-${sortOrder}`
-      };
-    });
-
     if (isFileExtended) {
       return (
         <div className="dir-tool">
@@ -142,7 +128,7 @@ class DirTool extends React.Component {
       <React.Fragment>
         <div className="dir-tool d-flex">
           <ViewModes currentViewMode={currentMode} switchViewMode={this.props.switchViewMode} />
-          <ReposSortMenu sortOptions={sortOptions} onSelectSortOption={this.onSelectSortOption}/>
+          <SortMenu sortBy={sortBy} sortOrder={sortOrder} onSelectSortOption={this.onSelectSortOption} />
           {(!isCustomPermission) &&
             <div className="cur-view-path-btn" onClick={this.showDirentDetail}>
               <span className="sf3-font sf3-font-info" aria-label={propertiesText} title={propertiesText}></span>
