+++ frontend/src/css/layout.css
@@ -169,7 +169,7 @@
 }
 
 .repos-container {
-  margin-bottom: 10rem;
+  padding-bottom: 10rem;
 }
 
 .table-drop-active::before {
