+++ frontend/src/pages/shared-libs/item.js
@@ -2,183 +2,18 @@ import React, { Component, Fragment } from 'react';
 import { Dropdown, DropdownToggle, DropdownMenu, DropdownItem } from 'reactstrap';
 import PropTypes from 'prop-types';
 import dayjs from 'dayjs';
-import classNames from 'classnames';
 import relativeTime from 'dayjs/plugin/relativeTime';
-import cookie from 'react-cookies';
 import { Link, navigate } from '@gatsbyjs/reach-router';
 import { gettext, siteRoot, isPro } from '../../utils/constants';
 import { seafileAPI } from '../../utils/seafile-api';
 import { Utils } from '../../utils/utils';
 import toaster from '../../components/toast';
-import Repo from '../../models/repo';
-import Loading from '../../components/loading';
-import EmptyTip from '../../components/empty-tip';
-import LibsMobileThead from '../../components/libs-mobile-thead';
 import ModalPortal from '../../components/modal-portal';
 import ShareDialog from '../../components/dialog/share-dialog';
-import SortOptionsDialog from '../../components/dialog/sort-options';
 import RepoMonitoredIcon from '../../components/repo-monitored-icon';
 import { GRID_MODE, LIST_MODE } from '../../components/dir-view-mode/constants';
-import ContextMenu from '../../components/context-menu/context-menu';
-import { hideMenu, handleContextClick } from '../../components/context-menu/actions';
 
 dayjs.extend(relativeTime);
-
-class Content extends Component {
-
-  constructor(props) {
-    super(props);
-    this.state = {
-      isItemFreezed: false
-    };
-    this.libItems = [];
-  }
-
-  freezeItem = (freezed) => {
-    this.setState({
-      isItemFreezed: freezed
-    });
-  };
-
-  sortByName = (e) => {
-    e.preventDefault();
-    const sortBy = 'name';
-    const sortOrder = this.props.sortOrder == 'asc' ? 'desc' : 'asc';
-    this.props.sortItems(sortBy, sortOrder);
-  };
-
-  sortByTime = (e) => {
-    e.preventDefault();
-    const sortBy = 'time';
-    const sortOrder = this.props.sortOrder == 'asc' ? 'desc' : 'asc';
-    this.props.sortItems(sortBy, sortOrder);
-  };
-
-  sortBySize = (e) => {
-    e.preventDefault();
-    const sortBy = 'size';
-    const sortOrder = this.props.sortOrder == 'asc' ? 'desc' : 'asc';
-    this.props.sortItems(sortBy, sortOrder);
-  };
-
-  onContextMenu = (event, repo) => {
-    event.preventDefault();
-    const id = 'shared-libs-item-menu';
-    const menuList = Utils.getSharedLibsOperationList(repo);
-    handleContextClick(event, id, menuList, repo);
-  };
-
-  setLibItemRef = (index) => item => {
-    this.libItems[index] = item;
-  };
-
-  getLibIndex = (lib) => {
-    return this.props.items.findIndex(item => {
-      return item.repo_id === lib.repo_id;
-    });
-  };
-
-  onMenuItemClick = (operation, currentObject, event) => {
-    const index = this.getLibIndex(currentObject);
-    this.libItems[index].onMenuItemClick(operation, event);
-
-    hideMenu();
-  };
-
-  render() {
-    const { loading, errorMsg, items, sortBy, sortOrder, theadHidden, inAllLibs, currentViewMode } = this.props;
-
-    const emptyTip = inAllLibs ?
-      <p className={`libraries-empty-tip-in-${currentViewMode}-mode`}>{gettext('No shared libraries')}</p> :
-      <EmptyTip
-        title={gettext('No shared libraries')}
-        text={gettext('No libraries have been shared directly with you. A shared library can be shared with full or restricted permission. If you need access to a library owned by another user, ask the user to share the library with you.')}
-      >
-      </EmptyTip>;
-
-    if (loading) {
-      return <Loading />;
-    } else if (errorMsg) {
-      return <p className="error text-center">{errorMsg}</p>;
-    } else {
-      // sort
-      const sortByName = sortBy == 'name';
-      const sortByTime = sortBy == 'time';
-      const sortBySize = sortBy == 'size';
-      const sortIcon = sortOrder == 'asc' ? <span className="sf3-font sf3-font-down rotate-180 d-inline-block"></span> : <span className="sf3-font sf3-font-down"></span>;
-
-      const desktopThead = (
-        <thead>
-          <tr>
-            <th width="4%"></th>
-            <th width="3%"><span className="sr-only">{gettext('Library Type')}</span></th>
-            <th width="35%"><a className="d-block table-sort-op" href="#" onClick={this.sortByName}>{gettext('Name')} {sortByName && sortIcon}</a></th>
-            <th width="10%"><span className="sr-only">{gettext('Actions')}</span></th>
-            <th width="14%"><a className="d-block table-sort-op" href="#" onClick={this.sortBySize}>{gettext('Size')} {sortBySize && sortIcon}</a></th>
-            <th width="17%"><a className="d-block table-sort-op" href="#" onClick={this.sortByTime}>{gettext('Last Update')} {sortByTime && sortIcon}</a></th>
-            <th width="17%">{gettext('Owner')}</th>
-          </tr>
-        </thead>
-      );
-
-      const isDesktop = Utils.isDesktop();
-      const itemsContent = (
-        <>
-          {items.map((item, index) => {
-            return <Item
-              ref={this.setLibItemRef(index)}
-              key={index}
-              data={item}
-              isDesktop={isDesktop}
-              isItemFreezed={this.state.isItemFreezed}
-              freezeItem={this.freezeItem}
-              onMonitorRepo={this.props.onMonitorRepo}
-              currentViewMode={currentViewMode}
-              onContextMenu={this.onContextMenu}
-            />;
-          })}
-        </>
-      );
-      const content = currentViewMode == LIST_MODE ? (
-        <>
-          <table className={classNames({ 'repos-container': !inAllLibs }, { 'table-thead-hidden': !(isDesktop && !theadHidden) })}>
-            {isDesktop ? desktopThead : <LibsMobileThead inAllLibs={inAllLibs} />}
-            <tbody>
-              {itemsContent}
-            </tbody>
-          </table>
-        </>
-      ) : (
-        <div className="d-flex justify-content-between flex-wrap">
-          {itemsContent}
-        </div>
-      );
-
-      return items.length ? (
-        <>
-          {content}
-          <ContextMenu
-            id="shared-libs-item-menu"
-            onMenuItemClick={this.onMenuItemClick}
-          />
-        </>
-      ) : emptyTip;
-    }
-  }
-}
-
-Content.propTypes = {
-  currentViewMode: PropTypes.string,
-  inAllLibs: PropTypes.bool.isRequired,
-  theadHidden: PropTypes.bool.isRequired,
-  loading: PropTypes.bool.isRequired,
-  errorMsg: PropTypes.string.isRequired,
-  items: PropTypes.array.isRequired,
-  sortBy: PropTypes.string.isRequired,
-  sortOrder: PropTypes.string.isRequired,
-  sortItems: PropTypes.func.isRequired,
-  onMonitorRepo: PropTypes.func.isRequired
-};
 
 class Item extends Component {
 
@@ -530,141 +365,4 @@ Item.propTypes = {
   onContextMenu: PropTypes.func.isRequired,
 };
 
-class SharedLibraries extends Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-      loading: true,
-      errorMsg: '',
-      items: [],
-      sortBy: cookie.load('seafile-repo-dir-sort-by') || 'name', // 'name' or 'time' or 'size'
-      sortOrder: cookie.load('seafile-repo-dir-sort-order') || 'asc', // 'asc' or 'desc'
-      isSortOptionsDialogOpen: false
-    };
-  }
-
-  componentDidMount() {
-    if (!this.props.repoList) {
-      seafileAPI.listRepos({ type: 'shared' }).then((res) => {
-        let repoList = res.data.repos.map((item) => {
-          return new Repo(item);
-        });
-        this.setState({
-          loading: false,
-          items: Utils.sortRepos(repoList, this.state.sortBy, this.state.sortOrder)
-        });
-      }).catch((error) => {
-        this.setState({
-          loading: false,
-          errorMsg: Utils.getErrorMsg(error, true)
-        });
-      });
-    } else {
-      this.setState({
-        loading: false,
-        items: this.props.repoList
-      });
-    }
-  }
-
-  sortItems = (sortBy, sortOrder) => {
-    cookie.save('seafile-repo-dir-sort-by', sortBy);
-    cookie.save('seafile-repo-dir-sort-order', sortOrder);
-    this.setState({
-      sortBy: sortBy,
-      sortOrder: sortOrder,
-      items: Utils.sortRepos(this.state.items, sortBy, sortOrder)
-    });
-  };
-
-  toggleSortOptionsDialog = () => {
-    this.setState({
-      isSortOptionsDialogOpen: !this.state.isSortOptionsDialogOpen
-    });
-  };
-
-  onMonitorRepo = (repo, monitored) => {
-    let items = this.state.items.map(item => {
-      if (item.repo_id === repo.repo_id) {
-        item.monitored = monitored;
-      }
-      return item;
-    });
-    this.setState({ items: items });
-  };
-
-  renderContent = () => {
-    const { inAllLibs = false, currentViewMode = LIST_MODE, repoList } = this.props; // inAllLibs: in 'All Libs'('Files') page
-    const { items } = this.state;
-    return (
-      <Content
-        loading={this.state.loading}
-        errorMsg={this.state.errorMsg}
-        items={inAllLibs ? repoList : items}
-        sortBy={this.state.sortBy}
-        sortOrder={this.state.sortOrder}
-        sortItems={this.sortItems}
-        onMonitorRepo={this.onMonitorRepo}
-        theadHidden={inAllLibs}
-        inAllLibs={inAllLibs}
-        currentViewMode={currentViewMode}
-      />
-    );
-  };
-
-  renderSortIconInMobile = () => {
-    return (
-      <>
-        {(!Utils.isDesktop() && this.state.items.length > 0) && <span className="sf3-font sf3-font-sort action-icon" onClick={this.toggleSortOptionsDialog}></span>}
-      </>
-    );
-  };
-
-  render() {
-    const { inAllLibs = false, currentViewMode = LIST_MODE } = this.props; // inAllLibs: in 'All Libs'('Files') page
-    return (
-      <Fragment>
-        {inAllLibs ? (
-          <>
-            <div className={`d-flex justify-content-between mt-3 py-1 ${currentViewMode == LIST_MODE ? 'sf-border-bottom' : ''}`}>
-              <h4 className="sf-heading m-0">
-                <span className="sf3-font-share-with-me sf3-font nav-icon" aria-hidden="true"></span>
-                {gettext('Shared with me')}
-              </h4>
-              {this.renderSortIconInMobile()}
-            </div>
-            {this.renderContent()}
-          </>
-        ) : (
-          <div className="main-panel-center">
-            <div className="cur-view-container">
-              <div className="cur-view-path">
-                <h3 className="sf-heading m-0">{gettext('Shared with me')}</h3>
-                {this.renderSortIconInMobile()}
-              </div>
-              <div className="cur-view-content">
-                {this.renderContent()}
-              </div>
-            </div>
-          </div>
-        )}
-        {this.state.isSortOptionsDialogOpen &&
-        <SortOptionsDialog
-          toggleDialog={this.toggleSortOptionsDialog}
-          sortBy={this.state.sortBy}
-          sortOrder={this.state.sortOrder}
-          sortItems={this.sortItems}
-        />
-        }
-      </Fragment>
-    );
-  }
-}
-
-SharedLibraries.propTypes = {
-  currentViewMode: PropTypes.string,
-  inAllLibs: PropTypes.bool,
-  repoList: PropTypes.array,
-};
-
-export default SharedLibraries;
+export default Item;
